Update Session_Timeout.md
Updated the description for session timeout. The original description does not explain the usage of the term session but puts the worse case scenario of being attacked and stolen. From a end-user reading this, those whom do not have a clear understanding of the term and issue would identify the following comment as having session timeout means in the meantime anybody can attack and steal the session token. This could lead to drastic measures such as a constant request to refresh the token or an exceptionally short session token. This would not be effective in a real world scenario and as such i believe the change would highlight the definition of what a session is better.